+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Origins of Life Changelog
+## 26.2.0.2-alpha
+- Fossil blocks now mine correctly
+  - Pickaxe for deepslate and sulfur
+  - Shovel for clay
+- Adds more information to the mod screen
+
 ## 26.2.0.1-alpha
 - Adds some config options
 - Some new textures for